move autosell job
